quickfix: add the -Wconversion flag. then kept going and had to change some more things, just to make quickfix and everything used in it complete.

+++ src/nvim/option_defs.h
@@ -439,7 +439,7 @@ EXTERN long p_mmp;              /* 'maxmempattern' */
 EXTERN long p_mmt;              /* 'maxmemtot' */
 EXTERN long p_mis;              /* 'menuitems' */
 EXTERN char_u   *p_msm;         /* 'mkspellmem' */
-EXTERN long p_mls;              /* 'modelines' */
+EXTERN int p_mls;				/* 'modelines' */
 EXTERN char_u   *p_mouse;       /* 'mouse' */
 EXTERN char_u   *p_mousem;      /* 'mousemodel' */
 EXTERN long p_mouset;           /* 'mousetime' */
